Initial move of animation routines
